+++ test/sync_option_test.rb
@@ -35,4 +35,29 @@ class SyncOptionTest < MiniTest::Spec
     song.image.must_equal "processed via :sync: Module"
     song.band.name.must_equal "band, processed: Class"
   end
+
+
+  describe ":sync allows you conditionals" do
+    class BandForm < Reform::Form
+      property :name, sync: lambda { |value, options| options.user_options[:form].changed?(:name) ? model.name = value : nil } # change if it hasn't changed
+    end
+
+    let (:band) { Band.new("Metallica") }
+    let (:form) { BandForm.new(band) }
+
+    # don't set name, didn't change.
+    it do
+      band.instance_exec { def name=(*); raise; end }
+      form.validate("name" => "Metallica").must_equal true
+      form.sync
+      band.name.must_equal "Metallica"
+    end
+
+    # update name.
+    it do
+      form.validate("name" => "Iron Maiden").must_equal true
+      form.sync
+      form.name.must_equal "Iron Maiden"
+    end
+  end
 end